+++ src/argilla/tasks/users/__main__.py
@@ -18,7 +18,10 @@ from argilla.tasks.callback import init_callback
 
 from .create import create_user
 from .delete import delete_user
+from .list import list_users
 
 app = typer.Typer(help="Holds CLI commands for user management.", no_args_is_help=True, callback=init_callback)
+
 app.command(name="create", help="Creates a new user")(create_user)
 app.command(name="delete", help="Deletes a user")(delete_user)
+app.command(name="list", help="List users")(list_users)