+++ CHANGELOG.md
@@ -16,6 +16,16 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+### Added
+
+- Added `workspaces list` command to list Argilla workspaces ([#3594](https://github.com/argilla-io/argilla/pull/3594)).
+- Added `datasets list` command to list Argilla datasets ([#3658](https://github.com/argilla-io/argilla/pull/3658)).
+- Added `users create` command to create users ([#3667](https://github.com/argilla-io/argilla/pull/3667)).
+- Added `whoami` command to get current user ([#3673](https://github.com/argilla-io/argilla/pull/3673)).
+- Added `users delete` command to delete users ([#3671](https://github.com/argilla-io/argilla/pull/3671)).
+- Added `users list` command to list users ([#3688](https://github.com/argilla-io/argilla/pull/3688)).
+- Added `datasets push-to-hub` command to push a `FeedbackDataset` from Argilla into the HuggingFace Hub ([#3685](https://github.com/argilla-io/argilla/pull/3685)).
+
 ### Fixed
 
 - Fixed `Annotation queue` ([#3677](https://github.com/argilla-io/argilla/pull/3677))